requirements.txt
cftime
dask
intake
intake-esm
netcdf4
numpy
pandas
pytest
pyyaml
xarray
